big fix struct, start api

+++ pkg/db/migrations.go
@@ -7,19 +7,19 @@ import (
 
 const (
 	linksSchema = `CREATE TABLE IF NOT EXISTS links (
-			           id BIGSERIAL PRIMARY KEY,
+			           id SERIAL PRIMARY KEY,
 		        short_url VARCHAR(50) UNIQUE NOT NULL,
 		     original_url TEXT NOT NULL,
 		       created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 		        is_custom BOOLEAN NOT NULL DEFAULT FALSE,
-		     clicks_count INTEGER NOT NULL DEFAULT 0);
+		     clicks_count INT NOT NULL DEFAULT 0);
 			 
 			 CREATE INDEX IF NOT EXISTS idx_links_short_url ON links(short_url);
 		     CREATE INDEX IF NOT EXISTS idx_links_created_at ON links(created_at);`
 
 	analyticsSchema = `CREATE TABLE IF NOT EXISTS analytics (
-			               id BIGSERIAL PRIMARY KEY,
-			          link_id BIGINT NOT NULL REFERENCES links(id) ON DELETE CASCADE,
+			               id SERIAL PRIMARY KEY,
+			          link_id INT NOT NULL REFERENCES links(id) ON DELETE CASCADE,
 			      accessed_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 			       user_agent TEXT,
 			       ip_address INET,
